backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
python-dateutil==2.8.2
openpyxl==3.1.2
gunicorn==21.2.0
PyJWT==2.8.0
Werkzeug==2.3.7